[QOL-7939] move reporting controller functions to submodule
- only the plugin class should integrate them
